backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
fpdf2==2.8.1
prisma==0.11.0
python-dotenv==1.0.0
asyncpg==0.29.0
